deduplicate tests
