+++ build.sbt
@@ -36,7 +36,9 @@ lazy val modules: List[ProjectReference] = List(
   core,
   circe,
   doobie,
+  sql,
   doobie_2,
+  skunk,
   generic,
   demo
 )
@@ -87,16 +89,6 @@ lazy val circe = project
     )
   )
 
-lazy val sql = project
-  .in(file("modules/sql"))
-  .enablePlugins(AutomateHeaderPlugin)
-  .disablePlugins(RevolverPlugin)
-  .dependsOn(circe)
-  .settings(commonSettings)
-  .settings(
-    name := "gsp-graphql-sql",
-  )
-
 lazy val doobie = project
   .in(file("modules/doobie"))
   .enablePlugins(AutomateHeaderPlugin)
@@ -119,25 +111,36 @@ lazy val doobie = project
     )
   )
 
+lazy val sql = project
+  .in(file("modules/sql"))
+  .enablePlugins(AutomateHeaderPlugin)
+  .disablePlugins(RevolverPlugin)
+  .dependsOn(circe)
+  .settings(commonSettings)
+  .settings(
+    name := "gsp-graphql-sql",
+    libraryDependencies ++= Seq(
+      "org.typelevel"     %% "cats-effect"            % catsEffectVersion,
+      "io.chrisdavenport" %% "log4cats-slf4j"         % log4catsVersion,
+      "ch.qos.logback"    %  "logback-classic"        % logbackVersion % "test",
+      "com.dimafeng"      %% "testcontainers-scala-scalatest"  % testContainersVersion % "test",
+      "com.dimafeng"      %% "testcontainers-scala-postgresql" % testContainersVersion % "test",
+    )
+  )
+
 lazy val doobie_2 = project
   .in(file("modules/doobie-2"))
   .enablePlugins(AutomateHeaderPlugin)
   .disablePlugins(RevolverPlugin)
-  .dependsOn(sql, circe)
+  .dependsOn(sql % "test->test;compile->compile", circe)
   .settings(commonSettings)
   .settings(
     name := "gsp-graphql-doobie-2",
     fork in Test := true,
     parallelExecution in Test := false,
     libraryDependencies ++= Seq(
-      "org.typelevel"     %% "cats-effect"            % catsEffectVersion,
-      "io.chrisdavenport" %% "log4cats-slf4j"         % log4catsVersion,
       "org.tpolecat"      %% "doobie-core"            % doobieVersion,
       "org.tpolecat"      %% "doobie-postgres-circe"        % doobieVersion,
-
-      "ch.qos.logback"    %  "logback-classic"        % logbackVersion % "test",
-      "com.dimafeng"      %% "testcontainers-scala-scalatest"  % testContainersVersion % "test",
-      "com.dimafeng"      %% "testcontainers-scala-postgresql" % testContainersVersion % "test",
     )
   )
 
@@ -145,21 +148,15 @@ lazy val skunk = project
   .in(file("modules/skunk"))
   .enablePlugins(AutomateHeaderPlugin)
   .disablePlugins(RevolverPlugin)
-  .dependsOn(sql, circe)
+  .dependsOn(sql % "test->test;compile->compile", circe)
   .settings(commonSettings)
   .settings(
     name := "gsp-graphql-skunk",
     fork in Test := true,
     parallelExecution in Test := false,
     libraryDependencies ++= Seq(
-      "org.typelevel"     %% "cats-effect"            % catsEffectVersion,
-      "io.chrisdavenport" %% "log4cats-slf4j"         % log4catsVersion,
       "org.tpolecat"      %% "skunk-core"            % "0.0.21",
       "org.tpolecat"      %% "skunk-circe"        % "0.0.21",
-
-      "ch.qos.logback"    %  "logback-classic"        % logbackVersion % "test",
-      "com.dimafeng"      %% "testcontainers-scala-scalatest"  % testContainersVersion % "test",
-      "com.dimafeng"      %% "testcontainers-scala-postgresql" % testContainersVersion % "test",
     )
   )
