--- modules/skunk/src/test/resources/logback-test.xml
@@ -1,15 +1,0 @@
-<configuration>
-  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-    <encoder>
-      <pattern>[%thread] %highlight(%-5level) %cyan(%logger{15}) - %msg %n</pattern>
-    </encoder>
-  </appender>
-
-  <root level="WARN">
-    <appender-ref ref="STDOUT" />
-  </root>
-
-  <logger name="utils.DatabaseSuite" level="INFO" additivity="false">
-    <appender-ref ref="STDOUT" />
-  </logger>
-</configuration>
