+++ modules/sql/src/main/scala/Row.scala
@@ -1,3 +1,6 @@
+// Copyright (c) 2016-2020 Association of Universities for Research in Astronomy, Inc. (AURA)
+// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
 package edu.gemini.grackle.sql
 
 /** A database row is a list of untyped values (some of which may be `FailedJoin`). */
