+++ modules/sql/src/test/resources/db/coalesce.sql
