+++ modules/sql/src/test/resources/db/embedding.sql
